Kali: Fixed merge conflict and realigned directory structure
