+++ frontend/src/pages/ForgotPassword/index.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 
-import ForgetImage from '../assets/img/login-banner.png'
+import ForgetImage from '../../assets/img/login-banner.png'
+
 const ForgetPassword = () => {
     return (
         <div className="content">
@@ -9,11 +10,11 @@ const ForgetPassword = () => {
                 <div className="row">
                     <div className="col-md-8 offset-md-2">
 
-                         {/*Account Content*/}
+                        {/*Account Content*/}
                         <div className="account-content">
                             <div className="row align-items-center justify-content-center">
                                 <div className="col-md-7 col-lg-6 login-left">
-                                    <img src={ForgetImage} className="img-fluid" alt="Login Banner" />
+                                    <img src={ForgetImage} className="img-fluid" alt="Login Banner"/>
                                 </div>
                                 <div className="col-md-12 col-lg-6 login-right">
                                     <div className="login-header">
@@ -22,11 +23,11 @@ const ForgetPassword = () => {
                                             link</p>
                                     </div>
 
-                                     {/*Forgot Password Form */}
+                                    {/*Forgot Password Form */}
                                     <form action="https://dreamguys.co.in/demo/doccure/login.html">
                                         <div className="form-group form-focus">
-                                            <input type="email" className="form-control floating" />
-                                                <label className="focus-label">Email</label>
+                                            <input type="email" className="form-control floating"/>
+                                            <label className="focus-label">Email</label>
                                         </div>
                                         <div className="text-right">
                                             <a className="forgot-link" href="login.html">Remember your password?</a>
@@ -35,7 +36,7 @@ const ForgetPassword = () => {
                                                 type="submit">Reset Password
                                         </button>
                                     </form>
-                                     {/*/Forgot Password Form */}
+                                    {/*/Forgot Password Form */}
 
                                 </div>
                             </div>
